// src/enum/log/category.enum.ts
export enum LogCategoryEnum
{
    TransactionFailed = 'transaction_failed',
    TransactionSuccess = 'transaction_success',
    TransactionPending = 'transaction_pending',
    TransactionCanceled = 'transaction_canceled',
    TransactionError= 'transaction_error',
    PaymentDoneWithoutConfirmation='payment_done_without_confirmation'
}